fix: moved model versions to a shared service
