+++ src/ui/pages/Settings.tsx
@@ -26,58 +26,9 @@ import {
   Typography,
 } from '@mui/material';
 import { LlmValidationService } from '@service/llmValidation';
+import { MODEL_VERSIONS } from '@service/modelVersions';
 import { useViewModel } from '@state/hook/useViewModel';
 import React, { useCallback, useEffect, useId, useState } from 'react';
-
-type GptVersion = 'gpt-4o' | 'gpt-4o-mini' | 'gpt-4.1' | 'o1-mini' | 'o3' | 'o4-mini';
-type ClaudeVersion = 'claude-3-5-haiku-latest' | 'claude-3-5-sonnet-latest' | 'claude-3-7-sonnet-latest';
-type GeminiVersion = 'gemini-2.0-flash' | 'gemini-2.0-flash-lite' | 'gemini-2.5-flash-preview-04-17' | 'gemini-2.5-pro-preview-05-06';
-
-interface ModelConfig<T extends LlmVersion> {
-  default: T;
-  options: readonly T[];
-  labels: Record<T, string>;
-}
-
-interface ModelVersions {
-  GPT: ModelConfig<GptVersion>;
-  CLAUDE: ModelConfig<ClaudeVersion>;
-  GEMINI: ModelConfig<GeminiVersion>;
-}
-
-export const MODEL_VERSIONS: ModelVersions = {
-  GPT: {
-    default: 'gpt-4o',
-    options: ['gpt-4o', 'gpt-4o-mini', 'gpt-4.1', 'o1-mini', 'o3', 'o4-mini'] as const,
-    labels: {
-      'gpt-4o': 'GPT-4o',
-      'gpt-4o-mini': 'GPT-4o Mini',
-      'gpt-4.1': 'GPT-4.1',
-      'o1-mini': 'o1-mini',
-      'o3': 'o3',
-      'o4-mini': 'o4-mini',
-    },
-  },
-  CLAUDE: {
-    default: 'claude-3-7-sonnet-latest',
-    options: ['claude-3-5-haiku-latest', 'claude-3-5-sonnet-latest', 'claude-3-7-sonnet-latest'] as const,
-    labels: {
-      'claude-3-5-haiku-latest': 'Claude 3.5 Haiku',
-      'claude-3-5-sonnet-latest': 'Claude 3.5 Sonnet',
-      'claude-3-7-sonnet-latest': 'Claude 3.7 Sonnet',
-    },
-  },
-  GEMINI: {
-    default: 'gemini-2.0-flash',
-    options: ['gemini-2.0-flash', 'gemini-2.0-flash-lite', 'gemini-2.5-flash-preview-04-17', 'gemini-2.5-pro-preview-05-06'] as const,
-    labels: {
-      'gemini-2.0-flash': 'Gemini 2.0 Flash',
-      'gemini-2.0-flash-lite': 'Gemini 2.0 Flash Lite',
-      'gemini-2.5-flash-preview-04-17': 'Gemini 2.5 Flash Preview',
-      'gemini-2.5-pro-preview-05-06': 'Gemini 2.5 Pro Preview',
-    },
-  },
-};
 
 function getValidVersion(modelKey: Llm, currentVersion: string | undefined): LlmVersion {
   const config = MODEL_VERSIONS[modelKey];
@@ -226,16 +177,16 @@ const LlmModelSettingRow: React.FC<LlmModelSettingRowProps> = ({
                       >
                         {isValidating
                           ? (
-                            <CircularProgress size={20} />
-                          )
+                              <CircularProgress size={20} />
+                            )
                           : isValid === true
                             ? (
-                              <CheckIcon color="success" />
-                            )
+                                <CheckIcon color="success" />
+                              )
                             : isValid === false
                               ? (
-                                <ErrorIcon color="error" />
-                              )
+                                  <ErrorIcon color="error" />
+                                )
                               : null}
                       </div>
                     </InputAdornment>
